+++ src/main/java/com/xiliulou/electricity/entity/TenantNoteRecharge.java
@@ -38,7 +38,7 @@ public class TenantNoteRecharge {
     /**
      * 短信充值次数
      */
-    private Integer rechargeNum;
+    private Long rechargeNum;
     
     /**
      * 充值时间
